added spacing, reorganized cfiles
