+++ BottomTabs.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { useTheme } from 'react-native-paper';
-import HomeScreen from './screens/HomeScreen';
-import UploadScreen from './screens/UploadScreen';
-import FormsScreen from './screens/FormsScreen';
-import ScheduleScreen from './screens/ScheduleScreen';
+import HomeScreen from './screens/Home/HomeScreen';
+import UploadScreen from './screens/Documents/UploadScreen';
+import FormsScreen from './screens/Forms/FormsScreen';
+import ScheduleScreen from './screens/Schedule/ScheduleScreen';
 
 const Tab = createMaterialBottomTabNavigator();
