+++ screens/Documents/UploadScreen.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Text, View } from 'react-native';
-import { screenStyles } from './HomeScreen';
+import styles from '../ScreenStyles';
 
 const UploadScreen = () => {
   return (
-    <View style={screenStyles.text}>
-      <Text>Upload your necessary documents!</Text>
+    <View style={styles.root}>
+      <Text style={styles.text}>Upload your necessary documents!</Text>
     </View>
   );
 };
