+++ styles/GlobalSpacing.tsx
@@ -1,0 +1,8 @@
+export default {
+  FONT_SMALL: 16,
+  FONT_MEDIUM: 20,
+  FONT_LARGE: 24,
+  PADDING: 16,
+  FONT_WEIGHT_NORMAL: 'normal',
+  FONT_WEIGHT_BOLD: 'bold',
+};
